+++ global-client/src/logic/store/messagebox.ts
@@ -1,2 +1,2 @@
-import { iframeHubId, miniClientId, MessageType, Message, MessageBoxType, useMessageBox } from '../../../../hub-client/src/logic/store/messagebox';
+import { iframeHubId, miniClientId, MessageType, Message, MessageBoxType, useMessageBox } from '../../../../hub-client/src/logic/store/messagebox.js';
 export { iframeHubId, miniClientId, MessageType, Message, MessageBoxType, useMessageBox };