+++ hub-client/src/logic/composables/useMatrixFiles.ts
@@ -1,32 +1,11 @@
 import { usePubHubs } from '@/logic/core/pubhubsStore';
+import { imageTypes, mediaTypes, fileTypes, allTypes } from '@/model/constants';
 
 const useMatrixFiles = () => {
 	const pubhubs = usePubHubs();
 	const downloadUrl = pubhubs.getBaseUrl + '/_matrix/media/r0/download/';
 	const uploadUrl = pubhubs.getBaseUrl + '/_matrix/media/r0/upload';
 	const deleteUrl = pubhubs.getBaseUrl + '/_synapse/admin/v1/media/';
-
-	const imageTypes = ['image/png', 'image/jpeg', 'image/gif', 'image/svg'];
-	const mediaTypes = ['audio/wave', 'audio/wav', 'audio/x-wav', 'audio/x-pn-wav', 'audio/webm', 'video/webm', 'audio/ogg', 'video/ogg', 'application/ogg'];
-	const fileTypes = [
-		'application/pdf',
-		'application/txt',
-		'text/plain',
-		'application/vnd.oasis.opendocument.presentation',
-		'application/vnd.oasis.opendocument.text',
-		'application/vnd.ms-powerpoint',
-		'application/vnd.openxmlformats-officedocument.presentationml.presentation',
-		'application/rtf',
-		'application/vnd.ms-excel',
-		'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-		'application/zip',
-		'text/calendar',
-		'application/msword',
-		'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-		'text/csv',
-	];
-
-	const allTypes = [...imageTypes, ...mediaTypes, ...fileTypes];
 
 	function getTypesAsString(types: Array<string>) {
 		return types.join(',');