+++ hub-client/src/model/constants.ts
@@ -22,4 +22,27 @@ enum OnboardingType {
 	full = 'full',
 }
 
-export { RelationType, RoomEmit, OnboardingType };
+// File types
+const imageTypes = ['image/png', 'image/jpeg', 'image/gif', 'image/svg'];
+const mediaTypes = ['audio/wave', 'audio/wav', 'audio/x-wav', 'audio/x-pn-wav', 'audio/webm', 'video/webm', 'audio/ogg', 'video/ogg', 'application/ogg'];
+const fileTypes = [
+	'application/pdf',
+	'application/txt',
+	'text/plain',
+	'application/vnd.oasis.opendocument.presentation',
+	'application/vnd.oasis.opendocument.text',
+	'application/vnd.ms-powerpoint',
+	'application/vnd.openxmlformats-officedocument.presentationml.presentation',
+	'application/rtf',
+	'application/vnd.ms-excel',
+	'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+	'application/zip',
+	'text/calendar',
+	'application/msword',
+	'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+	'text/csv',
+];
+
+const allTypes = [...imageTypes, ...mediaTypes, ...fileTypes];
+
+export { RelationType, RoomEmit, OnboardingType, imageTypes, mediaTypes, fileTypes, allTypes };