+++ hub-client/src/registerPlugins.js
@@ -4,38 +4,31 @@
  *
  */
 
-const requirePlugin = require.context(
-	'./plugins',
-	// Whether or not to look in subfolders
-	true,
-	// The regular expression used to match plugin filenames
-	/plugin.*\.(js|ts)$/,
-);
-
 export const registerPlugins = (app) => {
-	// register all the plugins in a global variable
+	const pluginFiles = import.meta.glob('./plugins/plugin.*.{js,ts}');
+
 	app.config.globalProperties._plugins = [];
-	requirePlugin.keys().forEach((fileName) => {
-		let pluginConfig = requirePlugin(fileName);
-		pluginConfig.plugin._path = fileName.replace(/\.\/(.*?)\/.*/g, '$1'); // Transform name, to path. Skipping filename part.
-		app.config.globalProperties._plugins.push(pluginConfig.plugin);
-	});
+
+	for (const path in pluginFiles) {
+		pluginFiles[path]().then((pluginModule) => {
+			const pluginConfig = pluginModule.plugin;
+			pluginConfig._path = path.replace(/\.\/plugins\/(.*?)\/.*/g, '$1'); // Transform name to path
+			app.config.globalProperties._plugins.push(pluginConfig);
+		});
+	}
 };
 
-const requirePluginComponent = require.context(
-	'./plugins',
-	// Whether or not to look in subfolders
-	true,
-	// The regular expression used to match base component filenames
-	/Plugin[A-Z]\w+\.vue$/,
-);
+export const registerPluginComponents = (app) => {
+	const componentFiles = import.meta.glob('./plugins/Plugin[A-Z]*.vue');
 
-export const registerPluginComponents = (app) =>
-	requirePluginComponent.keys().forEach((fileName) => {
-		const componentConfig = requirePluginComponent(fileName);
-		const componentName = fileName
-			.split('/')
-			.pop()
-			.replace(/\.\w+$/, '');
-		app.component(componentName, componentConfig.default || componentConfig);
-	});
+	for (const path in componentFiles) {
+		componentFiles[path]().then((componentModule) => {
+			const componentName =
+				path
+					.split('/')
+					.pop()
+					?.replace(/\.\w+$/, '') || '';
+			app.component(componentName, componentModule.default || componentModule);
+		});
+	}
+};