+++ CHANGELOG.md
@@ -10,6 +10,9 @@ _Please add a brief description of any changes and any migrations to be performe
  * _(Use the [MIGRATE] and [BREAKING] prefixes together with another one if that makes more sense.)_
 
 
+- [UPDATE] Replaced vue.cli by Vite
+- [NEW] Message can be added to a image or file in the chat
+- [NEW] Added a markdown editor for hub description, house rules, and contact information
 - [NEW] List of the members of a room can be shown now (button in header of the room)
 
 ## 12 June 2025 - v2.1.3